+++ src/main/java/com/project/snackpick/controller/member/MemberViewController.java
@@ -1,4 +1,4 @@
-package com.project.snackpick.controller;
+package com.project.snackpick.controller.member;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
